+++ HISTORY.md
@@ -4,3 +4,5 @@ History
 ## 1.0.0 / 2015-03-11
 
 *  init
+*  + Date.prototype.format
+*  + Date.prototype.toLocaleString